+++ backend/src/main/java/com/wellnessapp/dto/mealpost/MealPostResponse.java
@@ -1,16 +1,17 @@
 package com.wellnessapp.dto.mealpost;
 
 import java.time.Instant;
+import java.math.BigDecimal;
 
 public record MealPostResponse(
         Long id,
         Long plannedMealId,
         String mealType,
         String mealName,
-        int calories,
-        int proteinGrams,
-        int carbsGrams,
-        int fatGrams,
+        BigDecimal calories,
+        BigDecimal proteinGrams,
+        BigDecimal carbsGrams,
+        BigDecimal fatGrams,
         Instant postedAt,
         String imageUrl,
         String clientRequestId
